Move MavenUrl to a separate file
Subclasses should be placed into separate files, not inlined into other
files. Some small code cleanups in MavenUrl as well.
